# Improve NetworkPolicy batch installation