@tnqn 2021-07-29
> LGTM overall. I assume this would supercede #2465?

@Dyanngg Yes

> 
> > I do wonder if all the differences in the code between initial batch installation and regular
> > ongoing rule processing are completely necessary. In both cases, it seems that we should be computing a diff, and applying the required operations as one bundle.
> 
> My understanding is that in case of batch rule install, we don't actually care about the intermediate conjunctive context changes caused by a single rule -- we only care about the final OVS flows, and in case of bundle failure, we simply revert the entire cache. This PR takes advantage of this fact and skips calculating those intermediate flow states, so the logic here is a bit different than async rule install.

@antoninbas Yes, the main motivation of batch install was to avoid priority reassigning but it can avoid intermediate flows for conjunctive match flow update as well. If rules are installed asynchrously:
1. Installing rule B may update priority of rule A's flows, and installing rule C may affect rule A and B.
2. Installing rule B may update the shared conjunctive match flow if it has same from/to address as A, and installing rule C may update it again.

This is what I got from ovs-vswitchd log:
With # v1.2.0, flows were calculated incrementally, 100136 modification are for conjunction match flow update:
```
2021-07-27T07:09:05.226Z|00197|connmgr|INFO|br-int<->unix#0: 159657 flow_mods in the 9 s starting 10 s ago (59520 adds, 1 deletes, 100136 modifications)
```

Based on v1.2.0, if we simply remove batch install, the number of modification flows is same as above but took more time to install. If the rules have different priorties, there would be more modification flows caused by priority-reassigning:
```

2021-07-27T07:10:59.359Z|00139|connmgr|INFO|br-int<->unix#0: 133058 flow_mods in the 9 s starting 10 s ago (59230 adds, 1 deletes, 73827 modifications)
2021-07-27T07:11:59.359Z|00140|connmgr|INFO|br-int<->unix#0: 51223 flow_mods in the 59 s starting 60 s ago (24913 adds, 1 deletes, 26309 modifications)
```

With this patch, there is no modification flow and all NetworkPolicy flows were installed via single bundle.
```
2021-07-27T11:23:07.830Z|00139|connmgr|INFO|br-int<->unix#0: 61759 flow_mods in the last 10 s (61758 adds, 1 deletes)
```