@tnqn 2021-07-30
> I understand the rationale for this change, or for treating the restart differently from regular rule updates, which is not new to this PR. But ultimately it seems to me that we are just reconciling between a current state and a desired state. In the restart case the current state is `EMPTY`, and we can compute the desired state from all the received rules, and do the actual reconciliation in one go using only `ADD` operations. In the case of ongoing rule updates, the diff between current & desired state is typically small, and requires a combination of `ADD` / `UPDATE` / `DELETE` operations. But ultimately, we do reconciliation between 2 states using an OpenFlow bundle. So my question was more future-looking, and I was wondering about whether we could represent the state in such a fashion that the logic is _mostly_ the same in both cases, without sacrificing performance. That may not be possible though...

Good suggestion. I actually thought a little about this when I was working on this change and tried to unify the single rule installation and multiple rules installation but stopped when the change was much bigger than needed for this bug itself. I will continue thinking about your suggestion and try to simplify/unify the approach as follow-up.